Remove LayerRenderPass
