+++ packages/render-composer/src/RenderPipeline.tsx
@@ -1,6 +1,7 @@
 import { useNullableState } from "@hmans/use-nullable-state"
+import { useThree } from "@react-three/fiber"
 import * as PP from "postprocessing"
-import React, { createContext, useContext } from "react"
+import React, { createContext, useContext, useMemo } from "react"
 import * as RC from "."
 
 const RenderPipelineContext = createContext<{
@@ -19,21 +20,29 @@ export const RenderPipeline = ({
   children,
   transparentFXLayer = 2
 }: RenderPipelineProps) => {
+  const camera = useThree((s) => s.camera)
   const [depthCopyPass, setDepthCopyPass] = useNullableState<PP.DepthCopyPass>()
   const [copyPass, setCopyPass] = useNullableState<PP.CopyPass>()
+
+  const cameraLayerMask = useMemo(() => camera.layers.mask, [camera])
 
   return (
     <RC.EffectComposer>
       {/* Render all scene objects _except_ for those on the transparent FX layer. */}
-      <RC.LayerRenderPass layerMask={~(1 << transparentFXLayer)} />
+      <RC.LambdaPass
+        fun={() => (camera.layers.mask = ~(1 << transparentFXLayer))}
+      />
+      <RC.RenderPass />
+      <RC.LambdaPass fun={() => (camera.layers.mask = cameraLayerMask)} />
       <RC.DepthCopyPass ref={setDepthCopyPass} />
       <RC.CopyPass ref={setCopyPass} />
 
       {/* Render just the transparent FX objects. */}
-      <RC.LayerRenderPass
-        layerMask={1 << transparentFXLayer}
-        ignoreBackground
+      <RC.LambdaPass
+        fun={() => (camera.layers.mask = 1 << transparentFXLayer)}
       />
+      <RC.RenderPass ignoreBackground />
+      <RC.LambdaPass fun={() => (camera.layers.mask = cameraLayerMask)} />
 
       {depthCopyPass && copyPass && (
         <RenderPipelineContext.Provider
