--- packages/render-composer/src/passes/LayerRenderPass.tsx
@@ -1,64 +1,0 @@
-import { useThree } from "@react-three/fiber"
-import { RenderPass } from "postprocessing"
-import { useContext, useLayoutEffect, useMemo } from "react"
-import {
-  Camera,
-  Material,
-  Scene,
-  WebGLRenderer,
-  WebGLRenderTarget
-} from "three"
-import { EffectComposerContext } from "../EffectComposer"
-
-export type LayerRenderPassProps = {
-  layerMask: number
-  ignoreBackground?: boolean
-  clear?: boolean
-}
-
-export const LayerRenderPass = ({
-  layerMask,
-  ignoreBackground = false,
-  clear = false
-}: LayerRenderPassProps) => {
-  const scene = useThree((s) => s.scene)
-  const camera = useThree((s) => s.camera)
-
-  const pass = useMemo(
-    () => new LayerRenderPassImpl(scene, camera, undefined, layerMask),
-    [scene, camera]
-  )
-
-  useLayoutEffect(() => {
-    pass.ignoreBackground = ignoreBackground
-    pass.clearPass.enabled = clear
-  })
-
-  useContext(EffectComposerContext).useItem(pass)
-
-  return null
-}
-
-export class LayerRenderPassImpl extends RenderPass {
-  constructor(
-    scene: Scene,
-    camera: Camera,
-    overrideMaterial?: Material,
-    public layerMask: number = 0xffffffff | 0
-  ) {
-    super(scene, camera, overrideMaterial)
-  }
-
-  render(
-    renderer: WebGLRenderer,
-    inputBuffer: WebGLRenderTarget,
-    outputBuffer: WebGLRenderTarget,
-    deltaTime: number,
-    stencilTest: boolean
-  ) {
-    const mask = this.camera.layers.mask
-    this.camera.layers.mask = this.layerMask
-    super.render(renderer, inputBuffer, outputBuffer, deltaTime, stencilTest)
-    this.camera.layers.mask = mask
-  }
-}
