+++ packages/render-composer/src/passes/RenderPass.tsx
@@ -1,16 +1,38 @@
 import { useThree } from "@react-three/fiber"
-import { RenderPass as RenderPassImpl } from "postprocessing"
-import { useContext, useMemo } from "react"
+import * as PP from "postprocessing"
+import {
+  forwardRef,
+  useContext,
+  useImperativeHandle,
+  useLayoutEffect,
+  useMemo
+} from "react"
 import { EffectComposerContext } from "../EffectComposer"
 
-export const RenderPass = () => {
-  const scene = useThree((s) => s.scene)
-  const camera = useThree((s) => s.camera)
-
-  const pass = useMemo(() => new RenderPassImpl(scene, camera), [scene, camera])
-
-  const { useItem } = useContext(EffectComposerContext)
-  useItem(pass)
-
-  return null
+export type RenderPassProps = {
+  ignoreBackground?: boolean
+  clear?: boolean
 }
+
+export const RenderPass = forwardRef<PP.RenderPass, RenderPassProps>(
+  ({ ignoreBackground = false, clear = false }, ref) => {
+    const scene = useThree((s) => s.scene)
+    const camera = useThree((s) => s.camera)
+
+    const pass = useMemo(
+      () => new PP.RenderPass(scene, camera),
+      [scene, camera]
+    )
+
+    useLayoutEffect(() => {
+      pass.ignoreBackground = ignoreBackground
+      pass.clearPass.enabled = clear
+    })
+
+    useImperativeHandle(ref, () => pass)
+
+    useContext(EffectComposerContext).useItem(pass)
+
+    return null
+  }
+)
