+++ packages/render-composer/src/passes/index.ts
@@ -2,5 +2,4 @@ export * from "./CopyPass"
 export * from "./DepthCopyPass"
 export * from "./EffectPass"
 export * from "./LambdaPass"
-export * from "./LayerRenderPass"
 export * from "./RenderPass"
